+++ packages/engine/src/utils/nitro-converter.ts
@@ -1,7 +1,7 @@
 import {Commitment, CommitmentType, Channel} from "fmg-core";
 import {CONSENSUS_LIBRARY_ADDRESS, NETWORK_ID, ETH_ASSET_HOLDER_ADDRESS, CHALLENGE_DURATION} from "../constants";
 import {appAttributesFromBytes, bytesFromAppAttributes} from "fmg-nitro-adjudicator/lib/consensus-app";
-import {bigNumberify, getAddress} from "ethers/utils";
+import {bigNumberify} from "ethers/utils";
 import {
   SignedState,
   State,
@@ -14,6 +14,7 @@ import {
 } from "@statechannels/nitro-protocol";
 import {SignedCommitment, signCommitment2} from "../domain/commitments";
 import {decodeConsensusData} from "@statechannels/nitro-protocol/lib/src/contract/consensus-data";
+import {convertAddressToBytes32, convertBytes32ToAddress} from "./data-type-utils";
 
 // This temporarily handles converting fmg-core entities to nitro-protocol entities
 // Eventually once nitro-protocol is more properly embedded in the engine this will go away
@@ -127,27 +128,6 @@ function convertToNitroChannel(channel: Channel): NitroChannel {
   };
 }
 
-export function convertBytes32ToAddress(bytes32: string): string {
-  const normalized = bigNumberify(bytes32).toHexString();
-  return getAddress(`0x${normalized.slice(-40)}`);
-}
-
-// e.g.,
-// 0x9546E319878D2ca7a21b481F873681DF344E0Df8 becomes
-// 0x0000000000000000000000009546E319878D2ca7a21b481F873681DF344E0Df8
-export function convertAddressToBytes32(address: string): string {
-  const normalizedAddress = bigNumberify(address).toHexString();
-  if (normalizedAddress.length !== 42) {
-    throw new Error(
-      `Address value is not right length. Expected length of 42 received length ${normalizedAddress.length} instead.`
-    );
-  }
-  // We pad to 66 = (32*2) + 2('0x')
-  return `0x${normalizedAddress
-    .toLowerCase()
-    .substr(2)
-    .padStart(64, "0")}`;
-}
 export function convertAllocationToOutcome({
   allocation,
   destination
